+++ src/cubes.c
@@ -12,7 +12,7 @@ static cube cubes[CUBES_NUM];
 
 void cube_dma_tiles(void)
 {
-	VDP_doVRamDMA(gfx_cubes, CUBE_VRAM_SLOT * 32 ,CUBE_VRAM_LEN * 16);
+	VDP_doVRamDMA((u32)gfx_cubes, CUBE_VRAM_SLOT * 32,CUBE_VRAM_LEN * 16);
 	// Cubes use Lyle's palette. No palette transfer here.
 }
